Add SetMaxBlocksPerScan

+++ common/consts/consts.go
@@ -5,6 +5,6 @@ import "time"
 const (
 	RetryInterval        = 3 * time.Second
 	DefaultMsgBuffer     = 1000
-	DefaultBlocksPerScan = 100
-	MaxBlocksPerScan     = 10000000
+	DefaultBlocksPerScan = uint64(100)
+	MaxBlocksPerScan     = uint64(10000000)
 )
